backend/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.38.0
packaging==21.3
pandas==1.5.3
Pillow==9.3.0
python-crontab==2.7.1
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
sqlparse==0.4.2
xlrd==2.0.1
xmltodict==0.13.0
fpdf2==2.6.1
django-shortuuidfield==0.1.3
googletrans==3.1.0a0
openpyxl==3.0.3
openpyxl-dictreader==0.1.9
zeep==4.2.1

# ХУР-тай холбоотой package
appdirs==1.4.4
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
defusedxml==0.7.1
idna==3.3
isodate==0.6.1
lxml==4.8.0
numpy==1.22.3
platformdirs==2.5.1
pycryptodome==3.14.1
pycryptodomex==3.14.1
requests-file==1.5.1
requests-toolbelt==0.9.1
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
urllib3==1.26.9
